Replace all u128 with u64 due to serde issue. 	modified:   CHANGELOG.md 	modified:   Cargo.toml 	modified:   examples/poise_example/src/main.rs 	modified:   src/model/http.rs 	modified:   src/model/search.rs 	modified:   src/model/track.rs 	modified:   src/player_context/context.rs 	modified:   src/player_context/mod.rs

+++ src/model/http.rs
@@ -22,7 +22,7 @@ pub struct UpdatePlayer {
     /// It must be a value above 0 or None.
     ///
     /// None resets this if it was set previously.
-    pub end_time: Option<u128>,
+    pub end_time: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     /// The player volume.
     ///
@@ -32,7 +32,7 @@ pub struct UpdatePlayer {
     /// The track position in milliseconds.
     /// 
     /// This value can be set to start a track at a specific time.
-    pub position: Option<u128>,
+    pub position: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     /// Whether the player should be paused.
     pub paused: Option<bool>,
